+++ src/editor/core/event/handlers/keydown.ts
@@ -1,4 +1,4 @@
-import { EditorZone } from '../../..'
+import { EditorMode, EditorZone } from '../../..'
 import { ZERO } from '../../../dataset/constant/Common'
 import { ElementType } from '../../../dataset/enum/Element'
 import { KeyMap } from '../../../dataset/enum/KeyMap'
@@ -24,11 +24,10 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
   const isCollapsed = startIndex === endIndex
   // 当前激活控件
   const control = draw.getControl()
-  const isPartRangeInControlOutside = control.isPartRangeInControlOutside()
   const activeControl = control.getActiveControl()
   if (evt.key === KeyMap.Backspace) {
-    if (isReadonly || isPartRangeInControlOutside) return
-    let curIndex: number
+    if (isReadonly || control.isPartRangeInControlOutside()) return
+    let curIndex: number | null
     if (activeControl) {
       curIndex = control.keydown(evt)
     } else {
@@ -58,14 +57,16 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
       }
       curIndex = isCollapsed ? index - 1 : startIndex
     }
+    if (curIndex === null) return
+    draw.getGlobalEvent().setCanvasEventAbility()
     rangeManager.setRange(curIndex, curIndex)
     draw.render({ curIndex })
   } else if (evt.key === KeyMap.Delete) {
-    if (isReadonly || isPartRangeInControlOutside) return
-    let curIndex: number
+    if (isReadonly || control.isPartRangeInControlOutside()) return
+    let curIndex: number | null
     if (activeControl) {
       curIndex = control.keydown(evt)
-    } else if (elementList[endIndex + 1]?.type === ElementType.CONTROL) {
+    } else if (elementList[endIndex + 1]?.controlId) {
       curIndex = control.removeControl(endIndex + 1)
     } else {
       if (!isCollapsed) {
@@ -79,10 +80,12 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
       }
       curIndex = isCollapsed ? index : startIndex
     }
+    if (curIndex === null) return
+    draw.getGlobalEvent().setCanvasEventAbility()
     rangeManager.setRange(curIndex, curIndex)
     draw.render({ curIndex })
   } else if (evt.key === KeyMap.Enter) {
-    if (isReadonly || isPartRangeInControlOutside) return
+    if (isReadonly || control.isPartRangeInControlOutside()) return
     const enterText: IElement = {
       value: ZERO
     }
@@ -125,17 +128,38 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
   } else if (evt.key === KeyMap.Left) {
     if (isReadonly) return
     if (index > 0) {
-      const curIndex = startIndex - 1
+      const cursorPosition = position.getCursorPosition()
+      // 单词整体移动
+      let moveCount = 1
+      if (isMod(evt)) {
+        const LETTER_REG = draw.getLetterReg()
+        // 起始位置
+        const moveStartIndex =
+          evt.shiftKey && !isCollapsed && startIndex === cursorPosition?.index
+            ? endIndex
+            : startIndex
+        if (LETTER_REG.test(elementList[moveStartIndex]?.value)) {
+          let i = moveStartIndex - 1
+          while (i > 0) {
+            const element = elementList[i]
+            if (!LETTER_REG.test(element.value)) {
+              break
+            }
+            moveCount++
+            i--
+          }
+        }
+      }
+      const curIndex = startIndex - moveCount
       // shift则缩放选区
       let anchorStartIndex = curIndex
       let anchorEndIndex = curIndex
-      const cursorPosition = position.getCursorPosition()
       if (evt.shiftKey && cursorPosition) {
         if (startIndex !== endIndex) {
           if (startIndex === cursorPosition.index) {
             // 减小选区
             anchorStartIndex = startIndex
-            anchorEndIndex = endIndex - 1
+            anchorEndIndex = endIndex - moveCount
           } else {
             anchorStartIndex = curIndex
             anchorEndIndex = endIndex
@@ -146,10 +170,10 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
       }
       if (!~anchorStartIndex || !~anchorEndIndex) return
       rangeManager.setRange(anchorStartIndex, anchorEndIndex)
-      const isCollapsed = anchorStartIndex === anchorEndIndex
+      const isAnchorCollapsed = anchorStartIndex === anchorEndIndex
       draw.render({
-        curIndex: isCollapsed ? anchorStartIndex : undefined,
-        isSetCursor: isCollapsed,
+        curIndex: isAnchorCollapsed ? anchorStartIndex : undefined,
+        isSetCursor: isAnchorCollapsed,
         isSubmitHistory: false,
         isCompute: false
       })
@@ -158,11 +182,32 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
   } else if (evt.key === KeyMap.Right) {
     if (isReadonly) return
     if (index < positionList.length) {
-      const curIndex = endIndex + 1
+      const cursorPosition = position.getCursorPosition()
+      let moveCount = 1
+      // 单词整体移动
+      if (isMod(evt)) {
+        const LETTER_REG = draw.getLetterReg()
+        // 起始位置
+        const moveStartIndex =
+          evt.shiftKey && !isCollapsed && startIndex === cursorPosition?.index
+            ? endIndex
+            : startIndex
+        if (LETTER_REG.test(elementList[moveStartIndex + 1]?.value)) {
+          let i = moveStartIndex + 2
+          while (i < elementList.length) {
+            const element = elementList[i]
+            if (!LETTER_REG.test(element.value)) {
+              break
+            }
+            moveCount++
+            i++
+          }
+        }
+      }
+      const curIndex = endIndex + moveCount
       // shift则缩放选区
       let anchorStartIndex = curIndex
       let anchorEndIndex = curIndex
-      const cursorPosition = position.getCursorPosition()
       if (evt.shiftKey && cursorPosition) {
         if (startIndex !== endIndex) {
           if (startIndex === cursorPosition.index) {
@@ -170,7 +215,7 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
             anchorStartIndex = startIndex
             anchorEndIndex = curIndex
           } else {
-            anchorStartIndex = startIndex + 1
+            anchorStartIndex = startIndex + moveCount
             anchorEndIndex = endIndex
           }
         } else {
@@ -185,10 +230,10 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
         return
       }
       rangeManager.setRange(anchorStartIndex, anchorEndIndex)
-      const isCollapsed = anchorStartIndex === anchorEndIndex
+      const isAnchorCollapsed = anchorStartIndex === anchorEndIndex
       draw.render({
-        curIndex: isCollapsed ? anchorStartIndex : undefined,
-        isSetCursor: isCollapsed,
+        curIndex: isAnchorCollapsed ? anchorStartIndex : undefined,
+        isSetCursor: isAnchorCollapsed,
         isSubmitHistory: false,
         isCompute: false
       })
@@ -307,11 +352,11 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
       direction: isUp ? MoveDirection.UP : MoveDirection.DOWN
     })
   } else if (isMod(evt) && evt.key === KeyMap.Z) {
-    if (isReadonly) return
+    if (isReadonly && draw.getMode() !== EditorMode.FORM) return
     historyManager.undo()
     evt.preventDefault()
   } else if (isMod(evt) && evt.key === KeyMap.Y) {
-    if (isReadonly) return
+    if (isReadonly && draw.getMode() !== EditorMode.FORM) return
     historyManager.redo()
     evt.preventDefault()
   } else if (isMod(evt) && evt.key === KeyMap.C) {
@@ -344,12 +389,12 @@ export function keydown(evt: KeyboardEvent, host: CanvasEvent) {
     }
     evt.preventDefault()
   } else if (evt.key === KeyMap.TAB) {
-    draw.insertElementList([
-      {
-        type: ElementType.TAB,
-        value: ''
-      }
-    ])
+    const tabElement: IElement = {
+      type: ElementType.TAB,
+      value: ''
+    }
+    formatElementContext(elementList, [tabElement], startIndex)
+    draw.insertElementList([tabElement])
     evt.preventDefault()
   }
 }