+++ src/editor/dataset/enum/table/Table.ts
@@ -3,3 +3,12 @@ export enum TableBorder {
   EMPTY = 'empty',
   EXTERNAL = 'external'
 }
+
+export enum TdBorder {
+  BOTTOM = 'bottom'
+}
+
+export enum TdSlash {
+  FORWARD = 'forward', // 正斜线 /
+  BACK = 'back' // 反斜线 \
+}