+++ src/editor/index.ts
@@ -44,7 +44,7 @@ import { defaultCursorOption } from './dataset/constant/Cursor'
 import { IPageNumber } from './interface/PageNumber'
 import { defaultPageNumberOption } from './dataset/constant/PageNumber'
 import { VerticalAlign } from './dataset/enum/VerticalAlign'
-import { TableBorder } from './dataset/enum/table/Table'
+import { TableBorder, TdBorder, TdSlash } from './dataset/enum/table/Table'
 import { IFooter } from './interface/Footer'
 import { defaultFooterOption } from './dataset/constant/Footer'
 import { MaxHeightRatio, NumberType } from './dataset/enum/Common'
@@ -65,6 +65,8 @@ import { IRangeStyle } from './interface/Listener'
 import { Override } from './core/override/Override'
 import { defaultPageBreakOption } from './dataset/constant/PageBreak'
 import { IPageBreak } from './interface/PageBreak'
+import { LETTER_CLASS } from './dataset/constant/Common'
+import { INTERNAL_CONTEXT_MENU_KEY } from './dataset/constant/ContextMenu'
 
 export default class Editor {
   public command: Command
@@ -139,6 +141,7 @@ export default class Editor {
       height: 1123,
       scale: 1,
       pageGap: 20,
+      backgroundColor: '#FFFFFF',
       underlineColor: '#000000',
       strikeoutColor: '#FF0000',
       rangeAlpha: 0.6,
@@ -154,8 +157,8 @@ export default class Editor {
       marginIndicatorColor: '#BABABA',
       margins: [100, 120, 100, 120],
       pageMode: PageMode.PAGING,
-      tdPadding: 5,
-      defaultTrMinHeight: 40,
+      tdPadding: [0, 5, 5, 5],
+      defaultTrMinHeight: 42,
       defaultColMinWidth: 40,
       defaultHyperlinkColor: '#0000FF',
       paperDirection: PaperDirection.VERTICAL,
@@ -164,6 +167,9 @@ export default class Editor {
       wordBreak: WordBreak.BREAK_WORD,
       printPixelRatio: 3,
       maskMargin: [0, 0, 0, 0],
+      letterClass: [LETTER_CLASS.ENGLISH],
+      contextMenuDisableKeys: [],
+      scrollContainerSelector: '',
       ...options,
       header: headerOptions,
       footer: footerOptions,
@@ -248,6 +254,9 @@ export default class Editor {
 
 // 对外对象
 export {
+  EDITOR_COMPONENT,
+  LETTER_CLASS,
+  INTERNAL_CONTEXT_MENU_KEY,
   Editor,
   RowFlex,
   VerticalAlign,
@@ -256,7 +265,6 @@ export {
   ElementType,
   ControlType,
   EditorComponent,
-  EDITOR_COMPONENT,
   PageMode,
   ImageDisplay,
   Command,
@@ -264,6 +272,8 @@ export {
   BlockType,
   PaperDirection,
   TableBorder,
+  TdBorder,
+  TdSlash,
   MaxHeightRatio,
   NumberType,
   TitleLevel,