+++ src/editor/interface/Control.ts
@@ -1,6 +1,7 @@
 import { ControlType } from '../dataset/enum/Control'
 import { ICheckbox } from './Checkbox'
 import { IElement } from './Element'
+import { IRange } from './Range'
 
 export interface IValueSet {
   value: string
@@ -20,6 +21,10 @@ export interface IControlCheckbox {
   checkbox?: ICheckbox
 }
 
+export interface IControlRule {
+  deletable?: boolean
+}
+
 export interface IControlBasic {
   type: ControlType
   value: IElement[] | null
@@ -27,9 +32,13 @@ export interface IControlBasic {
   conceptId?: string
   prefix?: string
   postfix?: string
+  minWidth?: number
+  underline?: boolean
+  extension?: unknown
 }
 
 export type IControl = IControlBasic &
+  IControlRule &
   Partial<IControlSelect> &
   Partial<IControlCheckbox>
 
@@ -59,7 +68,31 @@ export interface IControlInstance {
 
   setValue(data: IElement[]): number
 
-  keydown(evt: KeyboardEvent): number
+  keydown(evt: KeyboardEvent): number | null
 
   cut(): number
+}
+
+export interface IControlContext {
+  range?: IRange
+  elementList?: IElement[]
+}
+
+export interface IGetControlValueOption {
+  conceptId: string
+}
+
+export type IGetControlValueResult = (Omit<IControl, 'value'> & {
+  value: string | null
+  innerText: string | null
+})[]
+
+export interface ISetControlValueOption {
+  conceptId: string
+  value: string
+}
+
+export interface ISetControlExtensionOption {
+  conceptId: string
+  extension: unknown
 }