+++ src/editor/interface/Editor.ts
@@ -6,6 +6,7 @@ import {
   WordBreak
 } from '../dataset/enum/Editor'
 import { ICheckboxOption } from './Checkbox'
+import { IPadding } from './Common'
 import { IControlOption } from './Control'
 import { ICursorOption } from './Cursor'
 import { IFooter } from './Footer'
@@ -38,6 +39,7 @@ export interface IEditorOption {
   height?: number
   scale?: number
   pageGap?: number
+  backgroundColor?: string
   underlineColor?: string
   strikeoutColor?: string
   rangeColor?: string
@@ -53,7 +55,7 @@ export interface IEditorOption {
   marginIndicatorColor?: string
   margins?: IMargin
   pageMode?: PageMode
-  tdPadding?: number
+  tdPadding?: IPadding
   defaultTrMinHeight?: number
   defaultColMinWidth?: number
   defaultHyperlinkColor?: string
@@ -62,6 +64,9 @@ export interface IEditorOption {
   historyMaxRecordCount?: number
   printPixelRatio?: number
   maskMargin?: IMargin
+  letterClass?: string[]
+  contextMenuDisableKeys?: string[]
+  scrollContainerSelector?: string
   wordBreak?: WordBreak
   header?: IHeader
   footer?: IFooter