+++ src/editor/utils/index.ts
@@ -216,14 +216,24 @@ export function cloneProperty<T>(
   }
 }
 
-export function omitObject<T>(object: T, keys: (keyof T)[]): T {
-  const cloneObject = deepClone(object)
+export function pickObject<T>(object: T, pickKeys: (keyof T)[]): T {
+  const newObject: T = <T>{}
   for (const key in object) {
-    if (keys.includes(key)) {
-      delete cloneObject[key]
+    if (pickKeys.includes(key)) {
+      newObject[key] = object[key]
     }
   }
-  return cloneObject
+  return newObject
+}
+
+export function omitObject<T>(object: T, omitKeys: (keyof T)[]): T {
+  const newObject: T = <T>{}
+  for (const key in object) {
+    if (!omitKeys.includes(key)) {
+      newObject[key] = object[key]
+    }
+  }
+  return newObject
 }
 
 export function convertStringToBase64(input: string) {